Added kjs -- The kernel javascript system.
This system allows the kernel debugger to be scripted with the new js
command.

Functions for reading and writing memory, as well as reading the module
list, the registry, and the trap frame registers are provided.  A simple
mechanism is provided for reading javascript initialization from the
registry so that the user may provide custom functions for use during
debugging sessions.

I have already used it to find a bug in kdb itself; as well as to list
sections in modules and perform various other tedious tasks often
required while debugging linked lists, etc.

This is a static library that will be built when KDBG=1, and linked
with ntoskrnl.


git-svn-id: file:///home/aleksey/repos/reactos_mirr/trunk@7554 9874b476-3cf2-4131-8f6d-dcf341c5f767