- Support for upper level filter drivers.
NOTE: Although the support for lower level filter driver is very similar, it's practically impossible to incorporate it with the current design without avoiding bugs such as #262 or #263. I don't want to extend the current mess beyond it's borders, so I ommited the call for initializing lower level filter drivers. The code needs a major cleanup and simplification on some places to make this support possible while maintaining a reasonable stable system.


git-svn-id: file:///home/aleksey/repos/reactos_mirr/trunk@8830 9874b476-3cf2-4131-8f6d-dcf341c5f767