Steven Edwards wanted me to commit this.
This adds a vew basic stubs of the SkyOS libraries libsky and libskygi. This actually was an experiment if SkyOS apps run in ReactOS/Win/WINE. It gets a slightly patched (only the optional PE header needs to be set to the right subsys code) version of TicTacToe released in the SkyOS forums. It doesn't crash but also doesn't work yet of course, but it successfully launches and calls the functions as you can see in the debug output.


git-svn-id: file:///srv/svn/reactos/trunk@10501 97493ccd-5924-5043-b1f5-66cb403b36ce