Turned on -Werror Propogate AdapterBinding rather than LogicalAdapter or MiniportBlock in places where we handle a request since we might need the protocol that did the request later on, get rid of LogicalAdapter completely as according to Filip, it's never needed.
Also, remove the Packet context hack in send complete if possible as this is
now unnecessary (we now send all needed information in the work item instead).

MiniQueueWorkItem, MiniDequeueWorkItem and MiniDoRequest were all changed to
take an AdapterBinding.
Added MiniRequestComplete, which handles a request complete.

Fixed hang in ipconfig and the tcpip control panel.


git-svn-id: file:///home/aleksey/repos/reactos_mirr/trunk@15429 9874b476-3cf2-4131-8f6d-dcf341c5f767