RosCMS update:
Update has been tested for 1 1/2 days and is already on our web-server. This commit don't effect ReactOS itself, as it is just for our website.

Updates:
* improved generator interface, more options, rarely used (debug) options removed
* complete new (from scratch) written page generator code; more than 5 times faster than the old one, plus new "update" feature
* new generator has a compatibility layer to support the old generator API
* the new generator was meant to be better structured and more compact, as it turned out it became a bit more extensive although it is still highly readable, as the code is divided into several helper functions.
* 875 code lines in comparision to 497 code lines of the old generator.

Information:
In real world usage, the new page generator is up to 20 times faster than the old one. This means that it takes now only seconds to generate all pages in comparion with the old generator (coded by me in sommer 2005) which generation time had to be counted in minutes for the same amount of pages.
Thanks to "Collibri" (from #reactos irc channel) who helped me with brainstorming and suggestions/ideas, the new generator code is now very efficient. The "generate everything" option is now O(log n) that means logarithmic in comparision to the old code which was O(n^c), polynomial.
The "generate altered pages" option has a O(n log n) linearithmic checking code and call for each page the above mentioned page generator (which is O(log n)).

Thanks to "Collibri" for help (see above for details).

git-svn-id: file:///home/aleksey/repos/reactos_mirr/trunk@24696 9874b476-3cf2-4131-8f6d-dcf341c5f767