ReactOS Translation Service:
I had worked on the ReactOS Translation Service (codename "ROST") in winter 2006/2007. It is almost ready, only one important part is missing, the user interface.
I decided to improve RosCMS to version 3, which comes now with an new user interface that speed up the whole work-flow a lot.
The plan is to use RosCMS v3 as flexible base for the translation service user interface. As RosCMS v3 is very flexible, both the interface and the database layout, it should be possible to implement it within a few days/weeks.

If someone is interested, the main parts of ROST are the parsers from XML => RC and XML => DB => XML. Magnus Olsen wrote an RC => XML parser in C which can be found on the ReactOS svn server too.
The main goal is to allow end user to translate ReactOS itself without have to mess around with RC-files. That means everyone (with language knowlegde) will be able to help to translate ReactOS.
The RC files syntax is a bit weird and outdated. We decided to parse the RC files and generate easy to handle XML files. The XML files are used behind the surface in ROST. An export interface make it possible to generate RC from the stored XML content. It is very important to keep the whole process lossless.
Magnus and I (Klemens) have tested ROST parsers on several RC files, from simple ones like NOTEPAD to complex ones like CMD and EXPLORER. Currently about 3/4 of the RC syntax is implemented and tested, although a bit work is needed to cover the whole syntax. Although, adding new syntax is straighforward (at least in the php based parsers).

Note: the current directory is a bit a mess, some scripts might be rearranged in future, etc.

git-svn-id: file:///srv/svn/reactos/trunk@26910 97493ccd-5924-5043-b1f5-66cb403b36ce