Revert hpoussin's commit in r29960 to unbreak build.
Most of *our* stuff has separated host/target headers now, but the Wine-shared components still include headers from include/reactos/wine.
And these headers (like "winnt.h" in that directory) include our headers from include/psdk themselves, which are meant to be used for target components, but not for host tools.

This needs to be fixed first, before we can remove this hack.

git-svn-id: file:///home/aleksey/repos/reactos_mirr/trunk@29964 9874b476-3cf2-4131-8f6d-dcf341c5f767