Revert the winddk.h changes of r32197 to make the hal compilable again without any warnings. Someone (preferably ros-arm-bringup ;-), please fix this correctly.
The warning is caused, because after the change in r32197, KeRaiseIrqlToSynchLevel is defined as NTKERNELAPI for all architectures.
NTKERNELAPI itself is defined to DECLSPEC_IMPORT if the header is not included from ntoskrnl. This is why hal outputs a warning for its KeRaiseIrqlToSynchLevel implementation.
As warnings are treated as errors for hal, the build fails.

git-svn-id: file:///home/aleksey/repos/reactos_mirr/trunk@32198 9874b476-3cf2-4131-8f6d-dcf341c5f767