[FASTFAT] Fix the most stupid bug ever in ReactOS history (at least!).
Dear Pierre, you pretend you're modifying the FileObject->FileName field, but, as you were told in the early when you learnt C, you're at that point in your code dealing with a local copy of the unicode string. So, if modifying the Buffer field is fine, because it's a pointer to memory zone, containing the wchar string, modifying the Length field is totally pointless because it won't magically modify the unicode string in the file object, and will just be forgotten as long the function will be unstacked.

So REALLY modify the file object. VfatSetRenameInformation receives less garbage information now...


git-svn-id: file:///srv/svn/reactos/trunk@63049 97493ccd-5924-5043-b1f5-66cb403b36ce