[NTFS] - Implement NtfsGetNameInformation() - Implement NtfsGetDirectoryInformation() - Implement NtfsGetFullDirectoryInformation() - Implement NtfsGetBothDirectoryInformation() - Implement NtfsQueryDirectory() which makes use of all the functions upper + all the previous fixes to make it able to look for a file using WC and file entry index
This is supposed to bring directory enumeration in line on NTFS volumes. BUT, there's a bug at some point which makes it fail (or my test volume is corrupted :-P (which is unlikely (Scheme, get out of this commit message!))). Will look into it later on. Committing so far because it's quite major work anyway.
More to follow~


git-svn-id: file:///srv/svn/reactos/trunk@64754 97493ccd-5924-5043-b1f5-66cb403b36ce