[FASTFAT] - Add a flag to enable (or disable depends on your reading ;-)) the VPB swapout process for volume umount. - By default, disable it for MSVC builds.
This *might* lolfix the MSVC builds.

Just for the record, we also have a bug in the GCC builds but it doesn't seem to show up. 
The main issue is that after formatting, we close handle to the volume, and IRP_MJ_CLEANUP is properly received and initiates a bit of dismount.
But then, the IRP_MJ_CLOSE is never received (why????) and thus, dismount cannot properly end.
This is a real & serious issue.


git-svn-id: file:///srv/svn/reactos/trunk@65527 97493ccd-5924-5043-b1f5-66cb403b36ce