[RTL] Finally, use the correct implementation for RtlAssert().
In testbots, the behavior will be the following:
On prompt, sysreg will always issue a 'break once' instruction. In umode, where kdbg isn't instructed to handle breakpoint, this will lead to application being killed. With a lot of noise. This matches previous behavior.
In kmode, where kdbg handles breakpoints, sysreg2 will ask for a bt and will reboot the VM. As it was done previously.

The testbots already have the newest available sysreg2 revision.

This was made possible thanks to the help of Stefan and Thomas.
It was also made possible thanks to the commits:
- in trunk: r68899, r68905, r68917
- in project-tools: r2074, r2232

CORE-10088 #resolve #comment Fixed with r68918


git-svn-id: file:///srv/svn/reactos/trunk@68918 97493ccd-5924-5043-b1f5-66cb403b36ce