[DOCUMENTATION] This is my written Bachelor Thesis called "Analysis, Design and Implementation of a Printing Stack for the Open-Source ReactOS Operating System". It's being uploaded in PDF and in the TeX source files.
This can serve as an introduction into the Windows Printing Stack and also detail some specifics of my implementation.
Additionally, the tools presented should be useful for newbies into Reverse Engineering.

Thanks to all the people who are keeping up ReactOS and thereby giving me the opportunity to make it part of my Bachelor Thesis!

git-svn-id: file:///srv/svn/reactos/trunk@69465 97493ccd-5924-5043-b1f5-66cb403b36ce