[BTRFS] Import the WinBtrfs 0.2 driver from https://github.com/maharmstone/btrfs.
Based on the initial work from Peter Hater, with various modification and patches sent upstream (yay, yet another collaboration :-)).

This driver is in its earlies, so expect crashes, issues, and so on. We'll keep it updated to get rid of these issues.
For now, it reads really well from a btrfs volume!

CORE-10892


git-svn-id: file:///srv/svn/reactos/trunk@71037 97493ccd-5924-5043-b1f5-66cb403b36ce