[RDBSS] [RXCE] [DDK] Add an initial implementation of the RXCE and RDBSS libraries that are used to implement mini-redirectors The associated headers are also added to DDK
The implementation is partial (really!) and is only working for read-only operations.

It leaks memory as hell (no refcounting implemented) and thus, objects freeing is not implemented.

It was tested with NFS41 driver. With such RDBSS, it's possible for the driver to mount a remote share, to list
files, to query their properties, to query volume properties, and finally to read files (with some corruption under
specific conditions).

Please refrain from committing in this (especially for modifying whitespaces or fixing comments...), this is still WIP
and under development, it would mess with my local changes!

In itself, it doesn't bring anything to ReactOS yet, as no mini redirector is available in ReactOS source tree.
This may come later on with NFS41.

More to follow...

CORE-11327


git-svn-id: file:///srv/svn/reactos/trunk@74674 97493ccd-5924-5043-b1f5-66cb403b36ce