[VGAFONTEDIT] Use the new header with SPDX license identifier and make the entire application GPL-2.0+
The code of the single GPL-2.0 only file "misc.c" was originally derived from GPL-2.0 only devmgmt.
I have now replaced it with the same code from LGPL-2.1+ mmc, which makes my application a full GPL-2.0+ work. This is legit, because I know that Thomas Weidenmueller is the original author of the "misc.c" code.

git-svn-id: file:///srv/svn/reactos/trunk@75987 97493ccd-5924-5043-b1f5-66cb403b36ce